Tidying up Memory Leaks
Still a few more to go

+++ Map.cpp
@@ -30,14 +30,13 @@ void Map::loadMap(std::string filename, sf::Vector2f pos) {
 					ssCell >> boardDims.x;
 					ssCell >> boardDims.y;
 					ssCell >> orientation;
-					row.push_back(Board(boardName, sf::Vector2i(pos.x + mapBounds.x * TILE_SOURCE_SIZE.x, pos.y + rowDim * TILE_SOURCE_SIZE.y), boardDims, orientation));
+					row.push_back(Board(boardName, sf::Vector2i((int)pos.x + mapBounds.x * TILE_SOURCE_SIZE.x, (int)pos.y + rowDim * TILE_SOURCE_SIZE.y), boardDims, orientation));
 				}
-				else
-				{
+				else { //  Empty Board
 					boardDims.x = map[boardPos.x - 1][boardPos.y].getDimensions().x;
 					boardDims.y = map[boardPos.x - 1][boardPos.y].getDimensions().y;
 					orientation = 0;
-					row.push_back(Board(sf::Vector2i(pos.x + mapBounds.x * TILE_SOURCE_SIZE.x, pos.y + rowDim * TILE_SOURCE_SIZE.y), boardDims, orientation));  //  Empty Board
+					row.push_back(Board(sf::Vector2i((int)pos.x + mapBounds.x * TILE_SOURCE_SIZE.x, (int)pos.y + rowDim * TILE_SOURCE_SIZE.y), boardDims, orientation));  
 				}
 
 				if (orientation == up || orientation == down)
@@ -47,8 +46,8 @@ void Map::loadMap(std::string filename, sf::Vector2f pos) {
 				boardPos.y++;
 				ssCell.str("");
 				ssCell.clear();
-
 			}
+
 			map.push_back(row);
 			if (orientation == up || orientation == down) 
 				mapBounds.x += boardDims.x;
